requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.9
django-ajax-datatable==4.3.1
django-bootstrap4==21.1
django-crispy-forms==1.13.0
django-widget-tweaks==1.4.9
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
#Pillow==8.4.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
